pyproject.toml
[project]
name = "sol-alive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyrogram>=2.0.106",
    "pyrsolace>=0.1.3",
    "python-dotenv>=1.0.1",
    "tgcrypto>=1.2.5",
]

[project.scripts]
main = "sol_alive.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.8",
]
